+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add module oracle_quota
 * Create a module utils ora_object which can contain classes to manage Oracle objects (ie. size clause).
 * Move inner classes from oracle_tablespace to ora_object.
+* Fix sanity tests on module oracle_services.
 
 ## 1.1.2 ##